+++ protected/controllers/TeachersController.php
@@ -227,12 +227,11 @@ class TeachersController extends Controller
                 $model->avatar = $_FILES['Teacher'];
                 if ($model->validate()) {
                     $src = Yii::getPathOfAlias('webroot') . "/images/teachers/" . $model->oldAvatar;
-                    if (is_file($src)) unlink($src);
                     $ext = substr(strrchr($_FILES['Teacher']['name']['foto_url'], '.'), 1);
                     $_FILES['Teacher']['name']['foto_url'] = uniqid() . '.' . $ext;
                     if (copy($_FILES['Teacher']['tmp_name']['foto_url'], Yii::getpathOfAlias('webroot') . "/images/teachers/" . $_FILES['Teacher']['name']['foto_url'])) {
                         $src = Yii::getPathOfAlias('webroot') . "/images/teacher/" . $model->oldAvatar;
-                        if (is_file($src))
+                        if (is_file($src) && $model->oldAvatar!='noname.png')
                             unlink($src);
                     }
                     $model->updateByPk($id, array('foto_url' => $_FILES['Teacher']['name']['foto_url']));
